# Faster parsing