Hi Jake,

I was on the fence on whether this is worth sending; for 50k inputs, which is about the max we'll realistically see, we're looking at 5-9ms gain. That's just a little bigger than noise. WDYT?

```
50k short flags
before: 11ms, 11ms, 13ms, 11ms, 17ms
after: 6ms, 5ms, 6ms, 5ms, 8ms

50k abbreviations
before: 97ms, 114ms, 120ms, 111ms, 107ms
after: 39ms, 37ms, 51ms, 36ms, 36ms

50k long flags
before: 27ms, 26ms, 24ms, 24ms, 28ms
after: 15ms, 19ms, 19ms, 15ms, 15ms

50k long options with =
before: 27ms, 26ms, 27ms, 24ms, 26ms
after: 15ms, 19ms, 19ms, 15ms, 15ms
```

Thanks.