@davidmorgan 2020-02-29
Thanks! Changelog updated. The pubspec currently says 1.5.3-dev, I think that's correct until we want to release? Please let me know :)